Updated SQL API for characters, ships, ranks to conform to new standard (return 404 instead of 200 OK when id not found)

+++ models/characters/model.js
@@ -8,12 +8,17 @@ function readAll(callback) {
 
 // get single
 function readOne(id, callback) {
-	database.all('SELECT characters.*, species.name AS "species.name" FROM characters JOIN species ON characters.species_id = species.id WHERE characters.id = ?', [id], callback);
+	database.get('SELECT characters.*, species.name AS "species.name" FROM characters JOIN species ON characters.species_id = species.id WHERE characters.id = ?', [id], callback);
 }
 
 // delete
 function remove(id, callback) {
-	database.run('DELETE FROM characters WHERE id = ?', [id], callback);
+	database.get('SELECT id FROM characters WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('DELETE FROM characters WHERE id = ?', [id], callback);
+		else
+			callback(404, null);
+	});
 }
 
 // create
@@ -36,14 +41,19 @@ function update(parameters, callback) {
 	});
 	paramFields.push(id);
 
-	database.run('UPDATE characters SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+	database.get('SELECT id FROM characters WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('UPDATE characters SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+		else
+			callback(404, null);
+	});
 }
 
 database.all('PRAGMA table_info(characters)', (error, rows) => {
-	if (error) console.error(error);
-	else {
+	if (error)
+		console.error(error);
+	else
 		rows.forEach(item => COLUMN_DATA.push(item));
-	}
 });
 
 module.exports = { readAll, readOne, create, update, remove, COLUMN_DATA };
