+++ models/ships/model.js
@@ -8,12 +8,17 @@ function readAll(callback) {
 
 // get single
 function readOne(id, callback) {
-	database.all('SELECT * FROM ships WHERE id = ?', [id], callback);
+	database.get('SELECT * FROM ships WHERE id = ?', [id], callback);
 }
 
 // delete
 function remove(id, callback) {
-	database.run('DELETE FROM ships WHERE id = ?', [id], callback);
+	database.get('SELECT id FROM ships WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('DELETE FROM ships WHERE id = ?', [id], callback);
+		else
+			callback(404, null);
+	});
 }
 
 // create
@@ -36,14 +41,19 @@ function update(parameters, callback) {
 	});
 	paramFields.push(id);
 
-	database.run('UPDATE ships SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+	database.get('SELECT id FROM ships WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('UPDATE ships SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+		else
+			callback(404, null);
+	});
 }
 
 database.all('PRAGMA table_info(ships)', (error, rows) => {
-	if (error) console.error(error);
-	else {
+	if (error)
+		console.error(error);
+	else
 		rows.forEach(item => COLUMN_DATA.push(item));
-	}
 });
 
 module.exports = { readAll, readOne, create, update, remove, COLUMN_DATA };
